# docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "BankAccountView" : {
        "properties" : {
          "accountName" : {
            "type" : "string"
          },
          "accountNumberMasked" : {
            "type" : "string"
          },
          "bankName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "DocumentUpload" : {
        "properties" : {
          "expiresOn" : {
            "description" : "Required for documents that expire (all except PROFILE_PHOTO)",
            "format" : "date",
            "type" : "string"
          },
          "file" : {
            "format" : "binary",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "NATIONAL_ID", "DRIVING_LICENSE", "PROFILE_PHOTO", "VEHICLE_REGISTRATION", "VEHICLE_INSURANCE" ],
            "type" : "string"
          },
          "vehicleId" : {
            "description" : "Required for VEHICLE_* documents: the vehicle assigned to you",
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "file", "type" ],
        "type" : "object"
      },
      "DocumentView" : {
        "properties" : {
          "contentType" : {
            "type" : "string"
          },
          "expiresOn" : {
            "format" : "date",
            "type" : "string"
          },
          "expiryFlaggedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rejectionReason" : {
            "type" : "string"
          },
          "reviewedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reviewedBy" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sizeBytes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING_REVIEW", "APPROVED", "REJECTED", "SUPERSEDED" ],
            "type" : "string"
          },
          "type" : {
            "enum" : [ "NATIONAL_ID", "DRIVING_LICENSE", "PROFILE_PHOTO", "VEHICLE_REGISTRATION", "VEHICLE_INSURANCE" ],
            "type" : "string"
          },
          "uploadedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "vehicleId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "DriverDetail" : {
        "properties" : {
          "bankAccount" : {
            "$ref" : "#/components/schemas/BankAccountView"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "documents" : {
            "items" : {
              "$ref" : "#/components/schemas/DocumentView"
            },
            "type" : "array"
          },
          "fullName" : {
            "type" : "string"
          },
          "history" : {
            "items" : {
              "$ref" : "#/components/schemas/HistoryView"
            },
            "type" : "array"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nationalIdMasked" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "readiness" : {
            "$ref" : "#/components/schemas/ReadinessView"
          },
          "status" : {
            "enum" : [ "PENDING", "DOCS_SUBMITTED", "TRAINING", "APPROVED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          },
          "suspension" : {
            "$ref" : "#/components/schemas/SuspensionView"
          },
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "vehicle" : {
            "$ref" : "#/components/schemas/VehicleView"
          }
        },
        "type" : "object"
      },
...
      "HistoryView" : {
        "properties" : {
          "actorUserId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "from" : {
            "enum" : [ "PENDING", "DOCS_SUBMITTED", "TRAINING", "APPROVED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          },
          "noticeAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "to" : {
            "enum" : [ "PENDING", "DOCS_SUBMITTED", "TRAINING", "APPROVED", "REJECTED", "SUSPENDED" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Problem" : {
        "additionalProperties" : true,
        "description" : "RFC 7807 problem details. Extra members (e.g. `missing`, `errors`) depend on `code`.",
        "properties" : {
          "code" : {
            "description" : "Stable error code for clients, e.g. `driver.documents-missing`",
            "example" : "driver.documents-missing"
          },
          "detail" : {
            "example" : "Some required documents have not been uploaded."
          },
          "instance" : { },
          "status" : {
            "example" : 409
          },
          "title" : {
            "example" : "Conflict"
          },
          "type" : {
            "example" : "about:blank"
          }
        }
      },
      "ReadinessView" : {
        "properties" : {
          "expired" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "hasActiveVehicle" : {
            "type" : "boolean"
          },
          "missing" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "pendingReview" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "RegisterVehicleRequest" : {
        "properties" : {
          "color" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "make" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "model" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "modelYear" : {
            "format" : "int32",
            "maximum" : 2100,
            "minimum" : 1980,
            "type" : "integer"
          },
          "plateNumber" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          },
          "seats" : {
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "vehicleClass" : {
            "enum" : [ "MOTO", "TUKTUK_REMORK", "TUKTUK_RICKSHAW", "CAR", "SUV", "VAN" ],
            "type" : "string"
          }
        },
        "required" : [ "plateNumber", "vehicleClass" ],
        "type" : "object"
      },
...
      "SuspensionView" : {
        "properties" : {
          "noticeAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "suspendedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "suspendedBy" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "VehicleView" : {
        "properties" : {
          "assignedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "make" : {
            "type" : "string"
          },
          "model" : {
            "type" : "string"
          },
          "modelYear" : {
            "format" : "int32",
            "type" : "integer"
          },
          "plateNumber" : {
            "type" : "string"
          },
          "seats" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "ACTIVE", "INACTIVE" ],
            "type" : "string"
          },
          "vehicleClass" : {
            "enum" : [ "MOTO", "TUKTUK_REMORK", "TUKTUK_RICKSHAW", "CAR", "SUV", "VAN" ],
            "type" : "string"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/v1/auth/otp/verify or /api/v1/auth/refresh",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Scheduled & contracted mobility platform for Phnom Penh.\n\n* **Authentication:** phone OTP (`/api/v1/auth/otp/request` → `/otp/verify`) returns a 15-minute JWT access token. Send it as `Authorization: Bearer <token>`. Refresh with `/api/v1/auth/refresh`.\n* **Retries:** mutating endpoints accept an `Idempotency-Key` header; a retry with the same key and body replays the original response (`Idempotent-Replayed: true`).\n* **Errors:** RFC 7807 `application/problem+json` with a stable machine-readable `code`; `title` and `detail` are localized from `Accept-Language` (`km` default, `en`).\n* **Time:** timestamps are UTC (ISO-8601); business dates are Asia/Phnom_Penh.",
    "title" : "Mobility API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/vehicles" : {
      "post" : {
        "description" : "The plate is normalized (uppercase, single spaces) and must be unique (`vehicle.plate-taken`).\n\n**Roles:** ADMIN",
        "operationId" : "register",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterVehicleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VehicleView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Register a vehicle",
        "tags" : [ "Admin · Vehicles" ],
        "x-required-roles" : [ "ADMIN" ]
      }
    },
    "/api/v1/admin/vehicles/{vehicleId}" : {
      "get" : {
        "description" : "**Roles:** ADMIN, DISPATCHER, SUPPORT, SAFETY_OFFICER",
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "vehicleId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VehicleView"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Not found"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Get a vehicle",
        "tags" : [ "Admin · Vehicles" ],
        "x-required-roles" : [ "ADMIN", "DISPATCHER", "SUPPORT", "SAFETY_OFFICER" ]
      }
    },
...
    "/api/v1/drivers/me" : {
      "get" : {
        "description" : "Status, documents, vehicle, `readiness` and history. Personal data is masked.\n\n**Roles:** DRIVER",
        "operationId" : "me_1",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverDetail"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "My driver profile",
        "tags" : [ "Driver app" ],
        "x-required-roles" : [ "DRIVER" ]
      }
    },
    "/api/v1/drivers/me/documents" : {
      "get" : {
        "description" : "**Roles:** DRIVER",
        "operationId" : "documents",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DocumentView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "My current documents",
        "tags" : [ "Driver app" ],
        "x-required-roles" : [ "DRIVER" ]
      },
      "post" : {
        "description" : "JPEG, PNG or PDF (checked by content), max 10 MB, encrypted at rest. Replaces the current document of the same type. PROFILE_PHOTO must be an image.\n\n**Roles:** DRIVER",
        "operationId" : "upload",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/DocumentUpload"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "413" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`document.too-large`: over 10 MB"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Upload a document",
        "tags" : [ "Driver app" ],
        "x-required-roles" : [ "DRIVER" ]
      }
    },
    "/api/v1/drivers/me/submit" : {
      "post" : {
        "description" : "PENDING → DOCS_SUBMITTED. Requires NATIONAL_ID, DRIVING_LICENSE and PROFILE_PHOTO (`driver.documents-missing` lists what is missing).\n\n**Roles:** DRIVER",
        "operationId" : "submit",
        "parameters" : [ {
          "description" : "Language of error messages and SMS: `km` (default) or `en`",
          "in" : "header",
          "name" : "Accept-Language",
          "required" : false,
          "schema" : {
            "enum" : [ "km", "en" ],
            "type" : "string"
          }
        }, {
          "description" : "Optional, 1–255 chars (a UUID is ideal). Retrying with the same key and body replays the first response with `Idempotent-Replayed: true`; a different body returns 422.",
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DriverDetail"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Invalid request (`code` explains why)"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Authenticated, but not allowed (role or scope)"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: invalid state transition, duplicate or concurrent update (`code` explains which)"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "`idempotency.key-reused`: same Idempotency-Key, different body"
          }
        },
        "security" : [ {
          "bearer-jwt" : [ ]
        } ],
        "summary" : "Submit documents for review",
        "tags" : [ "Driver app" ],
        "x-required-roles" : [ "DRIVER" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "This server",
    "url" : "/"
  } ],
...
}